Create LICENSE
Please review and accept this license in order to open source this repository.

Since Dodona is becoming more public, we migrated from github.ugent.be to github.com. We would like to eventually make Dodona open source, together with as much judges as possible. With this pull request we ask for your consent as contributor to publish your code under an MIT license.